+++ components/Comment/Comment.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState } from "react";
 
 import { deleteDoc, doc, updateDoc } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 import { useSession } from "next-auth/react";
 
 import { TiDeleteOutline } from "react-icons/ti";
@@ -15,7 +15,7 @@ interface CommentI {
   postId: string;
 }
 
-const Comment: FC<CommentI> = ({
+export const Comment: FC<CommentI> = ({
   userName,
   profileImg,
   text,
@@ -50,7 +50,7 @@ const Comment: FC<CommentI> = ({
     }
   };
   return (
-    <div className="flex justify-between items-center my-3">
+    <div className="flex justify-between items-center my-3 ">
       <div className="flex items-center">
         <div className="w-10 h-10">
           <img src={profileImg} alt="user" className="rounded-full " />
@@ -68,9 +68,9 @@ const Comment: FC<CommentI> = ({
             />
           </div>
         ) : (
-          <div className="flex">
+          <div className="flex items-center">
             <p className="ml-2 font-bold">{userName}</p>
-            <p className="ml-2">{text}</p>
+            <p className="ml-2 text-clip break-words max-w-[220px]">{text}</p>
           </div>
         )}
       </div>
@@ -89,5 +89,3 @@ const Comment: FC<CommentI> = ({
     </div>
   );
 };
-
-export default Comment;
