+++ components/Comment/index.ts
@@ -1,0 +1,1 @@
+export { Comment } from "./Comment";
