+++ components/CreatePost/index.ts
@@ -1,0 +1,1 @@
+export { CreatePost } from "./CreatePost";
