+++ components/Feed/index.ts
@@ -1,0 +1,1 @@
+export { Feed } from "./Feed";
