+++ components/LeftSideBar/index.ts
@@ -1,0 +1,1 @@
+export { LeftSideBar } from "./LeftSideBar";
