+++ components/RightSideBar/index.ts
@@ -1,0 +1,1 @@
+export { RightSideBar } from "./RightSideBar";
