+++ components/Skeleton/index.ts
@@ -1,0 +1,1 @@
+export { Skeleton } from "./Skeleton";
