+++ components/Stories/index.ts
@@ -1,0 +1,1 @@
+export { Stories } from "./Stories";
