+++ components/index.ts
@@ -1,0 +1,11 @@
+export { Comment } from "./Comment";
+export { CreatePost } from "./CreatePost";
+export { Feed } from "./Feed";
+export { Header } from "./Header";
+export { LeftSideBar } from "./LeftSideBar";
+export { Loader } from "./Loader";
+export { Post } from "./Post";
+export { Posts } from "./Posts";
+export { RightSideBar } from "./RightSideBar";
+export { Skeleton } from "./Skeleton";
+export { Stories } from "./Stories";
